Updating styling to Dark theme and Added Persistent Caching
Updating form to use a darm theme, as well as styled Project Details a bit better.
I think long run, the screen for updating the project details may just be build right
into editable fields in the main project page but for now it's working as is and looks
fine enough.

Also updated away from using styles.js since that's more of a react-native thing
and App.css is a better, cleaner way to do it in regular React.